@@ -1,12 +1,11 @@
 name: tagser
-description: A starting point for Dart libraries or applications.
-version: 1.0.1
+description: A small parser for XML-like documents. 
+version: 1.0.3
 homepage: https://github.com/Yohanson555/tagser4dart
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.14.2 <3.0.0'
 
 dev_dependencies:
-  pedantic: ^1.8.0
-  test: ^1.6.0
-  test_coverage: ^0.3.0
+  lints: ^1.0.0
+  test: ^1.18.2
