@@ -4,5 +4,8 @@ class TagserError {
   final int code;
   final String text;
 
-  TagserError({this.code, this.text});
+  TagserError({
+    required this.code,
+    required this.text,
+  });
 }
