@@ -1,9 +1,9 @@
 part of tagser;
 
 class TagserUtils {
-  static bool isAvailableCharacter(int charCode) {
-    if (charCode == CHAR_UNDERSCORE ||
-        (charCode >= 48 && charCode <= 57) ||
+  static bool isAvailableCharacter(int? charCode) {
+    if (charCode != null && charCode == charUnderscore ||
+        (charCode! >= 48 && charCode <= 57) ||
         (charCode >= 65 && charCode <= 90) ||
         (charCode >= 97 && charCode <= 122)) {
       return true;
