@@ -1,39 +1,29 @@
 library tagser;
 
 part 'src/attribute.dart';
-
 part 'src/const.dart';
-
 part 'src/context.dart';
-
 part 'src/error.dart';
-
 part 'src/messages.dart';
-
 part 'src/result.dart';
-
 part 'src/states.dart';
-
 part 'src/tag.dart';
-
 part 'src/utils.dart';
 
 class Tagser {
-  Map<String, dynamic> _options = {
+  final Map<String, dynamic> _options = {
     'ignoreCase': false,
   };
 
-  List<TagserState> _stack;
+  List<TagserState> _stack = [];
   int _line = 0; // template lining support
   int _symbol = 0;
 
-  Tagser({Map<String, dynamic> options}) {
-    if (options != null) {
-      _options.addAll(options);
-    }
+  Tagser({Map<String, dynamic>? options}) {
+    _options.addAll(options ?? {});
   }
 
-  List<Tag> parse(String source) {
+  List<Tag> parse(String? source) {
     _stack = [];
     _stack.add(RootState(null));
 
@@ -61,30 +51,24 @@ class Tagser {
 
           process(ProcessMessage(charCode: charCode), context);
         }
-
-        //if (l < lines.length - 1) {
-        //process(ProcessMessage(charCode: CHARS.ENTER));
-        //}
       }
 
-      process(ProcessMessage(charCode: CHAR_EOS), context);
+      process(ProcessMessage(charCode: charEos), context);
 
       final state = _stack.last;
 
-      if (_stack.length != 1 || !(state is RootState)) {
-        throw Exception('Source document malformed.');
-      } else {
-        return (state as RootState).tags;
+      if (state is RootState) {
+        return state.tags;
       }
     }
 
-    return null;
+    return [];
   }
 
   void process(TagserMessage msg, TagserContext context) {
     final state = _stack.last;
 
-    if (state != null && state.canAcceptMessage(msg)) {
+    if (state.canAcceptMessage(msg)) {
       final res = state.processMessage(msg, context);
 
       if (res != null) {
@@ -99,19 +83,15 @@ class Tagser {
     }
 
     if (r.state != null) {
-      _stack.add(r.state);
+      _stack.add(r.state!);
     }
 
     if (r.message != null) {
-      process(r.message, context);
+      process(r.message!, context);
     }
 
     if (r.err != null) {
-      final e = 'Error (${r.err.code}) on $currentLine:$_symbol ${r.err.text}';
-
-      print(e);
-
-      throw Exception(e);
+      throw Exception('Error (${r.err!.code}) on $currentLine:$_symbol ${r.err!.text}');
     }
   }
 
@@ -124,10 +104,46 @@ class Tagser {
   }
 
   setOption(String name, dynamic value) {
-    if (name != null && name.isNotEmpty) {
-      if (_options == null) _options = {};
+    _options[name] = value;
+  }
 
-      _options[name] = value;
+  String html(List<Tag>? tags) {
+    String res = '';
+
+    if (tags != null && tags.isNotEmpty) {
+      for (final tag in tags) {
+        if (tag.type == typeTag) {
+          res += '<${tag.name}';
+
+          // attrs
+
+          final attrs = tag.attributes;
+
+          if (attrs.isNotEmpty) {
+            attrs.forEach((name, attr) {
+              res += ' ${attr.name}="${attr.value}"';
+            });
+          }
+
+          res += '>';
+
+          // childs
+
+          final childs = tag.childs;
+
+          if (childs.isNotEmpty) {
+            res += html(childs);
+          }
+
+          //end
+
+          res += '</${tag.name}>';
+        } else {
+          res += tag.body != null ? tag.body! : '';
+        }
+      }
     }
+
+    return res;
   }
 }
